--- company/src/main/java/com/sibijo/company/dto/CompanyResponseDto.java
@@ -1,14 +1,0 @@
-package com.sibijo.company.dto;
-
-import lombok.*;
-import java.util.UUID;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class CompanyResponseDto {
-    private UUID hubId;
-    private String deliveryAddress;
-}
