--- company/src/main/java/com/sibijo/company/enums/CompanyType.java
@@ -1,6 +1,0 @@
-package com.sibijo.company.enums;
-
-public enum CompanyType {
-    생산업체,
-    수령업체
-}
