+++ company/src/main/java/com/sibijo/company/presentation/dto/CompanyRequest.java
@@ -1,0 +1,17 @@
+package com.sibijo.company.presentation.dto;
+
+import com.sibijo.company.domain.enums.CompanyType;
+import lombok.*;
+import java.util.UUID;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class CompanyRequest {
+    private String companyName;
+    private CompanyType companyType;  // "생산업체" 또는 "수령업체"
+    private UUID hubId;
+    private String address;
+}
