--- company/src/main/java/com/sibijo/company/repository/CompanyRepository.java
@@ -1,8 +1,0 @@
-package com.sibijo.company.repository;
-
-import com.sibijo.company.entity.Company;
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.UUID;
-
-public interface CompanyRepository extends JpaRepository<Company, UUID> {
-}
