--- product/src/main/java/com/sibijo/product/client/CompanyResponseDto.java
@@ -1,13 +1,0 @@
-package com.sibijo.product.client;
-
-import lombok.*;
-import java.util.UUID;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class CompanyResponseDto {
-    private UUID hubId;
-    private String deliveryAddress;
-}
