--- product/src/main/java/com/sibijo/product/controller/ProductController.java
@@ -1,54 +1,0 @@
-package com.sibijo.product.controller;
-
-import com.sibijo.product.dto.ProductRequest;
-import com.sibijo.product.dto.ProductResponseDto;
-import com.sibijo.product.entity.Product;
-import com.sibijo.product.service.ProductService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import java.util.List;
-import java.util.UUID;
-
-@RestController
-@RequestMapping("/api/products")
-@RequiredArgsConstructor
-public class ProductController {
-
-    private final ProductService productService;
-
-    @GetMapping
-    public ResponseEntity<List<Product>> getAllProducts() {
-        return ResponseEntity.ok(productService.getAllProducts());
-    }
-
-    @GetMapping("/{productId}")
-    public ResponseEntity<Product> getProduct(@PathVariable UUID productId) {
-        return ResponseEntity.ok(productService.getProductById(productId));
-    }
-
-    @PostMapping
-    public ResponseEntity<Product> createProduct(@RequestBody ProductRequest request) {
-        return ResponseEntity.ok(productService.createProduct(request));
-    }
-
-    @PutMapping("/{productId}")
-    public ResponseEntity<Product> updateProduct(@PathVariable UUID productId,
-            @RequestBody ProductRequest request) {
-        return ResponseEntity.ok(productService.updateProduct(productId, request));
-    }
-
-    @DeleteMapping("/{productId}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable UUID productId) {
-        productService.deleteProduct(productId);
-        return ResponseEntity.ok().build();
-    }
-
-    /**
-     * 주문 서비스에서 호출할 API – 상품의 재고 정보와 연결된 허브 ID 반환
-     */
-    @GetMapping("/{productId}/order")
-    public ResponseEntity<ProductResponseDto> getProductOrderInfo(@PathVariable UUID productId) {
-        return ResponseEntity.ok(productService.getProductOrderInfo(productId));
-    }
-}
