+++ product/src/main/java/com/sibijo/product/domain/repository/ProductRepository.java
@@ -1,0 +1,8 @@
+package com.sibijo.product.domain.repository;
+
+import com.sibijo.product.domain.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.UUID;
+
+public interface ProductRepository extends JpaRepository<Product, UUID> {
+}
