--- product/src/main/java/com/sibijo/product/dto/HubStockResponse.java
@@ -1,17 +1,0 @@
-package com.sibijo.product.dto;
-
-import lombok.*;
-import java.util.UUID;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class HubStockResponse {
-    private UUID hubStockId;
-    private UUID hubId;
-    private UUID companyId;
-    private UUID productId;
-    private Long amount;
-}
