--- product/src/main/java/com/sibijo/product/dto/ProductRequest.java
@@ -1,15 +1,0 @@
-package com.sibijo.product.dto;
-
-import lombok.*;
-import java.util.UUID;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ProductRequest {
-    private String productName;
-    private Integer price;
-    private UUID companyId;
-}
