--- product/src/main/java/com/sibijo/product/dto/ProductResponseDto.java
@@ -1,14 +1,0 @@
-package com.sibijo.product.dto;
-
-import lombok.*;
-import java.util.UUID;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ProductResponseDto {
-    private UUID hubId;
-    private Long amount;
-}
