--- product/src/main/java/com/sibijo/product/dto/UpdateStockRequest.java
@@ -1,13 +1,0 @@
-// UpdateStockRequest DTO
-package com.sibijo.product.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UpdateStockRequest {
-    private Long newAmount;
-}
