+++ product/src/main/java/com/sibijo/product/presentation/dto/HubStockResponse.java
@@ -1,0 +1,17 @@
+package com.sibijo.product.presentation.dto;
+
+import lombok.*;
+import java.util.UUID;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class HubStockResponse {
+    private UUID hubStockId;
+    private UUID hubId;
+    private UUID companyId;
+    private UUID productId;
+    private Long amount;
+}
