+++ product/src/main/java/com/sibijo/product/presentation/dto/ProductRequest.java
@@ -1,0 +1,15 @@
+package com.sibijo.product.presentation.dto;
+
+import lombok.*;
+import java.util.UUID;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ProductRequest {
+    private String productName;
+    private Integer price;
+    private UUID companyId;
+}
