+++ product/src/main/java/com/sibijo/product/presentation/dto/ProductResponseDto.java
@@ -1,0 +1,14 @@
+package com.sibijo.product.presentation.dto;
+
+import lombok.*;
+import java.util.UUID;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ProductResponseDto {
+    private UUID hubId;
+    private Long amount;
+}
