+++ product/src/main/java/com/sibijo/product/presentation/dto/UpdateStockRequest.java
@@ -1,0 +1,13 @@
+// UpdateStockRequest DTO
+package com.sibijo.product.presentation.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class UpdateStockRequest {
+    private Long newAmount;
+}
