--- product/src/main/java/com/sibijo/product/repository/ProductRepository.java
@@ -1,8 +1,0 @@
-package com.sibijo.product.repository;
-
-import com.sibijo.product.entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.UUID;
-
-public interface ProductRepository extends JpaRepository<Product, UUID> {
-}
